pyproject.toml
[project]
name = "polymath-v3"
version = "3.0.0"
description = "Polymathic research assistant - semantic search across scientific papers and code"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Max Van Belkum", email = "max.vanbelkum@vanderbilt.edu"}
]
keywords = ["rag", "semantic-search", "scientific-literature", "knowledge-graph"]

dependencies = [
    # Embeddings
    "FlagEmbedding>=1.2.0",
    "sentence-transformers>=2.2.0",

    # Database
    "psycopg[binary,pool]>=3.1.0",
    "pgvector>=0.2.0",
    "neo4j>=5.0.0",

    # PDF Processing
    "PyMuPDF>=1.23.0",
    "pdf2doi>=1.5.0",
    "pdfplumber>=0.10.0",

    # APIs
    "google-genai>=0.3.0",
    "google-cloud-batch>=0.10.0",
    "google-cloud-storage>=2.10.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",

    # Utilities
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "tqdm>=4.65.0",
    "rapidfuzz>=3.0.0",

    # MCP
    "mcp>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
polymath = "cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib", "cli", "mcp", "batch"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
